Documents/pwa-modern-stack/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo[srv]
